requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
# Editable install with no version control (boxinfo_crew==0.1.0)
-e c:\users\andys\documents\boxe agent ai\boxeinfo_crew\boxinfo_crew
brotli==1.2.0
browserbase==1.4.0
build==1.3.0
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.1
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
coverage==7.11.0
crewai==0.203.1
crewai-tools==0.76.0
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
ddgs==9.9.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
durationpy==0.10
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.1
fake-useragent==2.2.0
fastjsonschema==2.21.2
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
google-genai==1.45.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.11.3
invoke==2.2.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
kubernetes==34.1.0
lance-namespace==0.0.18
lance-namespace-urllib3-client==0.0.18
lancedb==0.25.2
langchain-anthropic==0.3.22
langchain-core==0.3.79
langchain-huggingface==0.3.1
langsmith==0.4.37
lark==1.3.0
litellm==1.74.9
llvmlite==0.44.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
minify_html==0.18.1
mistralai==1.9.10
mistune==3.1.4
mmh3==5.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
multimethod==1.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
notebook==7.4.7
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.3
oauthlib==3.3.1
onnxruntime==1.23.1
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
phik==0.12.5
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
portalocker==2.7.0
posthog==5.4.0
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyee==13.0.0
pyfiglet==1.0.4
pyflakes==3.4.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.38.2
PyMsgBox==1.0.9
pyparsing==3.2.5
pypdf==6.1.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytube==15.0.0
pytweening==1.2.0
pytz==2025.2
pyvis==0.3.2
PyWavelets==1.9.0
pywhatkit==5.4
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
stack-data==0.6.3
stagehand==0.5.4
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typeguard==4.4.4
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uv==0.9.3
uvicorn==0.37.0
visions==0.8.1
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wikipedia==1.4.0
wordcloud==1.9.4
yarl==1.22.0
ydata-profiling==4.17.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zstandard==0.25.0
